% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{check.FileDir}
\alias{check.FileDir}
\title{function to check the existance of the directory and combine directory and file names}
\usage{
check.FileDir(dirSave = NULL, fileName)
}
\arguments{
\item{dirSave}{specify a valid directory where to save the file}

\item{fileName}{name of the file}
}
\value{
directory and filename combined, ie path to fileName
}
\description{
function to check the existance of the directory and combine directory and file names
}
\keyword{internal}
