% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-utils.R
\name{comparison.Plt}
\alias{comparison.Plt}
\title{function that generates a comparison plot among several packages}
\usage{
comparison.Plt(
  pckgDwnlds.lst,
  t0,
  t1,
  cmb = TRUE,
  noMAvgs = FALSE,
  noCBs = FALSE,
  device = "PDF",
  dirSave = NULL
)
}
\arguments{
\item{pckgDwnlds.lst}{nested list containing the several packages to process}

\item{t0}{initial date}

\item{t1}{final date}

\item{cmb}{boolean flag, indicating whether the plots are shown in the same (default) graph or not}

\item{noMAvgs}{boolean flag, indicating whether moving averages are displayed (default) or NOT --set to TRUE--}

\item{noCBs}{boolean flag, indicating whether shaded confidence intervals are displayed (default) or NOT  --set to TRUE--}

\item{device}{string to select the output format: 'PDF'/'PNG'/'JPEG' or 'screen'}

\item{dirSave}{specify a valid directory where to save the file}
}
\description{
function that generates a comparison plot among several packages
}
\keyword{internal}
