% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadsStats-utils.R
\name{summaries}
\alias{summaries}
\title{function to display the summary of the data}
\usage{
summaries(data1, deltaTs = 30)
}
\arguments{
\item{data1}{first dataset, eg. total data}

\item{deltaTs}{a numerical (integer) value, indicating the lenght --in days-- for selecting a subset of the original dataset; default value is 1 mont, ie. 30 days}
}
\description{
function to display the summary of the data
}
\examples{
\donttest{
packageXdownloads <- retrievePckgData("ehelp")[[1]]
summaries(packageXdownloads)
}

}
